rpc-server/requirements.txt
web3==7.3.0
fastapi==0.115.0
python-dotenv==1.0.1
uvicorn==0.31.0
eth-account==0.13.4
